refactor(ui): rename hierarchy drawer files to modal paths
